[RCS]Set missing MSRP headers when sending a message.
Test: by manual
Bug: 179342541

+++ testapps/TestRcsApp/aosp_test_rcsclient/javatests/com/android/libraries/rcs/simpleclient/service/chat/SimpleChatSessionTest.java
@@ -52,7 +52,7 @@ public class SimpleChatSessionTest {
     SipSessionConfiguration configuration =
             new SipSessionConfiguration() {
                 @Override
-                public int getVersion() {
+                public long getVersion() {
                     return 0;
                 }
